refactor(admin): extract Basic Info and Input Mode sections from AppFormEditor (#1781)
Follow-up to #1781, continuing the series of extraction PRs already
opened against it (#1942, #2044, #2046, #2048, #2049, #2050, #2051,
#2054, #2060, #2061). These were the last two unclaimed sections per
the issue's tracking comments.

Extracts the Basic Information card (id/type/order/category/name/
description/color/icon/preferred model/allowed models/temperature/
output format/enabled/auto-start/ephemeral) into
BasicInfoSection.jsx, and the Input Mode & Microphone card (input
type, textarea rows, microphone settings, speech recognition
service) into InputModeSection.jsx, both under
client/src/features/admin/components/app-form/. Also moves the
shared parseNumberOrUndefined helper into
client/src/features/admin/utils/ since both the extracted sections
and the still-inline Upload/Transcription sections use it.

Pure extraction, no behavior change. AppFormEditor.jsx shrinks from
2,902 to 2,371 lines. Verified in a running dev server: edited the
Order field, toggled Input Type and the microphone checkbox
(confirming its conditional sub-fields show/hide correctly), zero
console errors.
